Addressing CI review comments.
