+++ tests/unit-tests/graphpartitioning/array_readwriter/__init__.py
@@ -1,0 +1,2 @@
+from . import csv, numpy_array, parquet
+from .registry import get_array_parser, register_array_parser
